# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Eventos API",
        "description": "Descripción de tu API",
        "contact": {
            "email": "roni290690@email.com"
        },
        "license": {
            "name": "Licencia",
            "url": ""
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/eventos/login": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar Usuarios",
                "description": "Descripción detallada de la ruta",
                "operationId": "geUsuarios",
                "requestBody": {
                    "description": "Datos para crear un nuevo ejemplo",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Usuario"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Login Obtener Token",
                "operationId": "authenticate",
                "requestBody": {
                    "description": "Request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Usuario"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/eventos/lista-detalle-juego": {
            "get": {
                "tags": [
                    "DetalleJuego"
                ],
                "summary": "Listar Detalle Juego",
                "description": "Reguest",
                "operationId": "listarDetalleJuego",
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/listar-detalle-juegoID/{idDetalleJuego}": {
            "get": {
                "tags": [
                    "DetalleJuego"
                ],
                "summary": "Listar Detalle Juego por ID",
                "operationId": "listarDetalleJuegoId",
                "parameters": [
                    {
                        "name": "idDetalleJuego",
                        "in": "path",
                        "description": "Id del Detalle de Juego",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "datos actualizado correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/create-detalle-juego": {
            "post": {
                "tags": [
                    "DetalleJuego"
                ],
                "summary": "Crear Detalle Juego",
                "operationId": "storeDetalleJuego",
                "requestBody": {
                    "description": "Request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DetalleJuego"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "datos actualizado correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/update-detalle-juego/{idDetalleJ}": {
            "put": {
                "tags": [
                    "DetalleJuego"
                ],
                "summary": "Actualizar Detalle Juego",
                "operationId": "updateDetalleJuego",
                "parameters": [
                    {
                        "name": "idDetalleJ",
                        "in": "path",
                        "description": "ID detalle Juego",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DetalleJuego"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "datos actualizado correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/eliminar-detalle-juego/{idDetalleJ}": {
            "put": {
                "tags": [
                    "DetalleJuego"
                ],
                "summary": "Eliminar Detalle Juego",
                "operationId": "destroyDetalleJuego",
                "parameters": [
                    {
                        "name": "idDetalleJ",
                        "in": "path",
                        "description": "ID detalle Juego",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "datos actualizado correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Usuario": {
                "required": [
                    "usuario",
                    "clave"
                ],
                "properties": {
                    "usuario": {
                        "type": "string",
                        "example": ""
                    },
                    "clave": {
                        "type": "string",
                        "example": ""
                    }
                },
                "type": "object"
            },
            "DetalleJuego": {
                "required": [
                    "idEvento",
                    "idJuego",
                    "fechIni",
                    "fechFin",
                    "idUsuario"
                ],
                "properties": {
                    "idEvento": {
                        "type": "integer"
                    },
                    "idJuego": {
                        "type": "integer"
                    },
                    "fechIni": {
                        "type": "string",
                        "format": "date"
                    },
                    "fechFin": {
                        "type": "string",
                        "format": "date"
                    },
                    "idUsuario": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}